Add internal tests
