+++ tests/test_sender.py
@@ -1,6 +1,7 @@
 from __future__ import print_function
 
 import os
+from re import I
 import pytest
 import six
 from six.moves import queue
@@ -438,3 +439,23 @@ def test_upgrade_removed(
 
 
 # TODO: test other sender methods
+
+
+@pytest.mark.parametrize("empty_query", [True, False])
+@pytest.mark.parametrize("local_none", [True, False])
+@pytest.mark.parametrize("oudated", [True, False])
+def test_exit_poll_local(
+    publish_util, mock_server, mocked_run, empty_query, local_none, oudated
+):
+    mock_server.ctx["out_of_date"] = oudated
+    mock_server.ctx["empty_query"] = empty_query
+    mock_server.ctx["local_none"] = local_none
+    publish_util()
+
+    out_of_date = mocked_run._poll_exit_response.local_info.out_of_date
+    if empty_query:
+        assert out_of_date
+    elif local_none:
+        assert not out_of_date
+    else:
+        assert out_of_date == oudated
